requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
Pillow==9.0.1
twmap==0.2.0
Werkzeug==2.0.3
